Alter: Move modify vs rebuild determination logic.
Now index alteration creates a pair of operations
(DropIndex + AddIndex) in any case, but during preparation
stage they could be replaced by one ModifyIndex operation.
This is bad by several reasons:

- Preparation stage is a traversal of a linked list of operations,
and the traversal is broken in such a replacement.

- A special order in the list of operation is required (drop
before add). It's also imposible to add drop of one index
followed by addition of another - the will be considered as
modification of the same index.

- Needless allocation and deallocation.

- dynamic_cast theoretically can be not supported.

Move the logic to on_replace_dd_index and create DropIndex+AddIndex
OR ModifyIndex appropriately.