Add MoveIndex and RebuildIndex operation
MoveIndex operation is used to move existing index from an old space to
a new one. RebuildIndex is used to recreate an index in a new space
instead of sequence of a DropIndex and an AddIndex.
All checks are gone from alter operations and prepare
operation purged. All checks are done in triggers before any alter
operation creating. Need for #2074