Add to bps_tree lower/upper_bound_elem methods
Originally bps_tree had only lower/upper_bound methods for
retreaving iterators by given key.

Sometimes it is necessary to find an iterator position by a value
that is (can be) stored in a tree.

For example memtx tree index holds a bps_tree with
(struct tuple *) values and (const char *) key.
lower/upper_bound methods used (const char *) as an argument,
but for stable iterators we need to find a position in the tree
by (struct tuple *) value.

Implement bps_tree methods that will allow to find an iterator
by a value (element) in the tree.

Needed for #1796