Fix memleak in iproto port, clean and optimize the code
Now if an error happens during the first tuple push from port to
obuf, the port is not actually destroyed - only the first tuple is
unreferenced.

I believe that port dump must not free any data at all, then once
a user creates a port he must take the ownership and destroy it in
the end of usage. That would simplify the user code, especially C++.

Simplify port_dump function and destroy a port in the code that
creates the port.